moved highlight pop up stuff to highlight folder, removed hardcoded login link, updated imports and snap
