+++ src/app/content/components/Toolbar/index.tsx
@@ -9,7 +9,7 @@ import { AppState, Dispatch } from '../../../types';
 import { assertDocument, assertString } from '../../../utils';
 import { clearSearch, openMobileToolbar, openSearchResultsMobile, requestSearch } from '../../search/actions';
 import * as selectSearch from '../../search/selectors';
-import HighlightButton from './MyHighlights/HighlightButton';
+import HighlightButton from './HighlightButton';
 import PrintButton from './PrintButton';
 import * as Styled from './styled';
 
@@ -111,8 +111,8 @@ class Toolbar extends React.Component<Props, State> {
             }
           </Styled.SearchInputWrapper>
         </Styled.SearchPrintWrapper>
-        <PrintButton />
         <HighlightButton/>
+        <PrintButton />
         <FormattedMessage id='i18n:toolbar:buy-book:text'>
           {(msg) => <Styled.BuyBookWrapper
             aria-label={msg}
